python_interface/requirements.txt
pyserial
requests
geocoder
PIL
